Fix some issues
Fix some issues

+++ FullStackPractice.Business/DepartmentService.cs
@@ -14,12 +14,15 @@ namespace FullStackPractice.Business
     public class DepartmentService : IDepartmentService
     {
         private readonly IUnitOfWork _unitOfWork;
-        private readonly AbstractValidator<Department> _departmentValidator;
+        private readonly UpdateDepartmentValidator _updateDepartmentValidator;
+        private readonly DeleteDepartmentValidator _deleteDepartmentValidator;
 
-        public DepartmentService(IUnitOfWork unitOfWork, AbstractValidator<Department> departmentValidator)
+        public DepartmentService(IUnitOfWork unitOfWork, UpdateDepartmentValidator updateDepartmentValidator,
+            DeleteDepartmentValidator deleteDepartmentValidator)
         {
             _unitOfWork = unitOfWork;
-            _departmentValidator = departmentValidator;
+            _updateDepartmentValidator = updateDepartmentValidator;
+            _deleteDepartmentValidator = deleteDepartmentValidator;
         }
 
         public async Task<List<Department>> GetAllDepartmentsAsync()
@@ -34,14 +37,14 @@ namespace FullStackPractice.Business
 
         public async Task CreateDepartmentAsync(Department department)
         {
-            var results = await _departmentValidator.ValidateAsync(department);
+            var results = await _updateDepartmentValidator.ValidateAsync(department);
             if (!results.IsValid)
             {
                 foreach (var failure in results.Errors)
                 {
                     throw new Exception(failure.ErrorMessage);
                 }
-            } 
+            }
             else
             {
                 await _unitOfWork.DepartmentRepository.AddAsync(department);
@@ -61,8 +64,19 @@ namespace FullStackPractice.Business
 
             if (department != null)
             {
-                await _unitOfWork.DepartmentRepository.RemoveAsync(department); 
-                await _unitOfWork.Complete();
+                var results = await _deleteDepartmentValidator.ValidateAsync(department);
+                if (!results.IsValid)
+                {
+                    foreach (var failure in results.Errors)
+                    {
+                        throw new Exception(failure.ErrorMessage);
+                    }
+                }
+                else
+                {
+                    await _unitOfWork.DepartmentRepository.RemoveAsync(department);
+                    await _unitOfWork.Complete();
+                }
             }
         }
     }
